Move inline JavaScript into dedicated asset files
This has a few benefits: Firstly, the JS is can be cached independently
of the page, decreasing total download size with a warm cache. Secondly,
a stricter CSP can be used that prohibits inline JavaScript, increasing
security. Lastly, this makes it possible to customize the JavaScript
without completely replacing the _Layout page.